requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1

